Implemented the ability to download a chapter's explanation for offline use
